Fix closed contour condition
If a cell has a saddle point, it is possible for a contour to return to
that cell without closing itself.  For example, a contour might leave a
cell through a N->E crossing, but eventually return through a W->S
crossing. This happens only if we start the contour trace from cell
types 5 or 10.  So we need to make the condition for a closed contour
more strict.  The contour not only has to return to its starting cell,
it must also end up at the correct edge.