Implement performance improvements using bit masking for crossing calculation
The chief suspect in allocation (and triggering GC) was all the small arrays allocated
to handle ambiguous cases. With this solution we use the 5th bit (0x10) to indicate
an ambiguous case. The logic to disambiguate is handled in the processing of crossings.
This significantly reduces memory allocations and yields modest performance improvements.